# internal/api/client/api/swagger.yaml
openapi: 3.0.3
info:
  title: Frank Auth Server API
  description: "Frank is a comprehensive authentication and authorization server that\
    \ provides OAuth2, passwordless login, MFA, \npasskeys, SSO, enterprise SSO, webhooks,\
    \ organizations, and API keys for machine-to-machine authentication.\n"
  contact:
    name: Frank Auth Server
    url: https://github.com/juicycleff/frank
  version: 1.0.0
servers:
- url: https://auth.example.com/api/v1
  description: Production server
- url: http://localhost:8080/api/v1
  description: Local development server
...
paths:
  /auth/login:
    post:
      tags:
      - Auth
      summary: Log in with email and password
      description: Authenticate a user with email and password
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/auth_login_body"
        required: true
      responses:
        "200":
          description: Successfully authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "403":
          description: Account locked or requires verification
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/register:
    post:
      tags:
      - Auth
      summary: Register a new user
      description: Create a new user account
      operationId: register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/auth_register_body"
        required: true
      responses:
        "201":
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /auth/refresh:
    post:
      tags:
      - Auth
      summary: Refresh access token
      description: Get a new access token using a refresh token
      operationId: refreshToken
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/auth_refresh_body"
        required: true
      responses:
        "200":
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_200_1"
        "401":
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/forgot-password:
    post:
      tags:
      - Auth
      summary: Request password reset
      description: Initiate the password reset process
      operationId: forgotPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/auth_forgotpassword_body"
        required: true
      responses:
        "202":
          description: Password reset email sent (if account exists)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_202"
  /auth/reset-password:
    post:
      tags:
      - Auth
      summary: Reset password
      description: Reset password using a token
      operationId: resetPassword
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/auth_resetpassword_body"
        required: true
      responses:
        "200":
          description: Password reset successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_200_2"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/verify-email:
    post:
      tags:
      - Auth
      summary: Verify email
      description: Verify email address using a token
      operationId: verifyEmail
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/auth_verifyemail_body"
        required: true
      responses:
        "200":
          description: Email verification successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/inline_response_200_3"
        "401":
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /users/me:
    get:
      tags:
      - Users
      summary: Get current user
      description: Get the profile of the currently logged in user
      operationId: getCurrentUser
      responses:
        "200":
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - bearerAuth: []
    put:
      tags:
      - Users
      summary: Update current user
      description: Update the profile of the currently logged in user
      operationId: updateCurrentUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/users_me_body"
        required: true
      responses:
        "200":
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      security:
      - bearerAuth: []
...
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          $ref: "#/components/schemas/Error_error"
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
          description: JWT access token
        refresh_token:
          type: string
        expires_at:
          type: integer
          format: int64
        mfa_required:
          type: boolean
        mfa_types:
          type: array
          items:
            type: string
      example:
        refresh_token: refresh_token
        expires_at: 0
        mfa_required: true
        user:
          metadata: ""
          last_login_at: 2000-01-23T04:56:07.000+00:00
          email_verified: true
          last_name: last_name
          active: true
          created_at: 2000-01-23T04:56:07.000+00:00
          profile_image_url: http://example.com/aeiou
          locale: locale
          updated_at: 2000-01-23T04:56:07.000+00:00
          phone_number: phone_number
          id: id
          first_name: first_name
          email: ""
        mfa_types:
        - mfa_types
        - mfa_types
        token: token
    User:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        first_name:
          type: string
        last_name:
          type: string
        phone_number:
          type: string
        profile_image_url:
          type: string
          format: uri
        locale:
          type: string
        email_verified:
          type: boolean
        active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        metadata:
          type: object
          additionalProperties: true
        last_login_at:
          type: string
          format: date-time
      example:
        metadata: ""
        last_login_at: 2000-01-23T04:56:07.000+00:00
        email_verified: true
        last_name: last_name
        active: true
        created_at: 2000-01-23T04:56:07.000+00:00
        profile_image_url: http://example.com/aeiou
        locale: locale
        updated_at: 2000-01-23T04:56:07.000+00:00
        phone_number: phone_number
        id: id
        first_name: first_name
        email: ""
...
    auth_login_body:
      required:
      - email
      - password
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        organization_id:
          type: string
        remember_me:
          type: boolean
          default: false
        captcha_response:
          type: string
    auth_register_body:
      required:
      - email
      - password
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        first_name:
          type: string
        last_name:
          type: string
        organization_id:
          type: string
        metadata:
          type: object
          additionalProperties: true
...
    auth_refresh_body:
      required:
      - refresh_token
      type: object
      properties:
        refresh_token:
          type: string
    inline_response_200_1:
      type: object
      properties:
        token:
          type: string
        refresh_token:
          type: string
        expires_at:
          type: integer
          format: int64
      example:
        refresh_token: refresh_token
        expires_at: 0
        token: token
    auth_forgotpassword_body:
      required:
      - email
      type: object
      properties:
        email:
          type: string
          format: email
    inline_response_202:
      type: object
      properties:
        message:
          type: string
          example: "If your email is registered, you will receive a password reset\
            \ link"
      example:
        message: "If your email is registered, you will receive a password reset link"
    auth_resetpassword_body:
      required:
      - new_password
      - token
      type: object
      properties:
        token:
          type: string
        new_password:
          type: string
          format: password
    inline_response_200_2:
      type: object
      properties:
        message:
          type: string
          example: Password has been reset successfully
      example:
        message: Password has been reset successfully
    auth_verifyemail_body:
      required:
      - token
      type: object
      properties:
        token:
          type: string
    inline_response_200_3:
      type: object
      properties:
        message:
          type: string
          example: Email verification successful
      example:
        message: Email verification successful
...
    users_me_body:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        phone_number:
          type: string
        profile_image_url:
          type: string
          format: uri
        locale:
          type: string
        metadata:
          type: object
          additionalProperties: true
...
    Error_error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        data:
          type: object
          additionalProperties: true
...
  securitySchemes:
    bearerAuth:
      type: http
      description: JWT token obtained from the /auth/login endpoint
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      description: API key for machine-to-machine authentication
      name: X-API-Key
      in: header